Refactoring structure and naming conventions
