+++ Client/.gitignore
@@ -19,6 +19,7 @@ dist
 .env.development.local
 .env.test.local
 .env.production.local
+/graveyard
 
 npm-debug.log*
 yarn-debug.log*
