+++ Client/package.json
@@ -8,11 +8,8 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@vitejs/plugin-react-swc": "^3.4.0",
-    "axios": "^1.5.0",
-    "prop-types": "^15.8.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.15.0",
     "vite-tsconfig-paths": "^4.2.1",
     "web-vitals": "^2.1.4"
   },
