+++ Client/src/App.jsx
@@ -1,14 +1,16 @@
 import Header from './components/Header/Header';
 import Home from './pages/Home';
-import { MinstaClone } from './components/InstaClone/InstaClone';
+import Project from './pages/Project';
+import Resources from './pages/Resources';
 
 function App() {
 
   return (
     <>
       <Header />
-      <Home />
-      {/* <MinstaClone /> */}
+      {/* <Home /> */}
+      {/* <Project /> */}
+      {/* <Resources /> */}
     </>
   );
 }
