+++ Client/src/assets/content/project.md
@@ -1,0 +1,40 @@
+# Clever Coders Wrangle Workaround for Meta’s Bogus Block
+
+## Jan 1, 2024
+
+TORONTO — ExtraExtra is stepping up for Canadians at a time when access to reliable information is more important than ever.
+
+Since August 2023, Meta has [blocked news][nyt] on Instagram and Facebook for users across Canada. The decision is a direct response to the Online News Act, a Federal bill requiring Big Tech to [share ad revenue][can] with news publishers. “This ban hurts all Canadians by placing barriers on high-quality journalism,” says ExtraExtra co-creator [Benny Goldman][linkedGold].
+
+ExtraExtra is a way to help Canadians stay informed. Follow [@ExtraExtraCA][insta] on Instagram to get the latest headlines. Visit [extraextranews.ca][website] to access full stories. News is curated from local, national, and international sources. According to lead developer [Benjamin Stubina][linkedStub], “we want to meet people where they're at while staying under Big Tech's radar.”
+
+To learn more about the implications of Meta's ban, visit the [Resources][resrc] page.
+
+<!-- {/* SocialLink, socialArray */} -->
+* [@ExtraExtraCA](https://www.instagram.com/ExtraExtraCA)
+* [/ExtraExtra](https://github.com/BenjaminStubina/ExtraExtra)
+* [E-mail us](mailto:extraextranewsca@gmail.com)
+<!-- {/* SocialLink, socialArray */} -->
+
+ExtraExtra was created in Toronto by Benjamin Stubina and Benny Goldman.
+
+<!-- {/* DevList, devArray */} -->
+[avatar] Benjamin Stubina [GitHub][gitStub], [LinkedIn][linkedStub] **optional: 1 sentence bio**
+<!-- {/* DevList, devArray */} -->
+[avatar] Benny Goldman [GitHub][gitGold], [LinkedIn][linkedGold]   **optional: 1 sentence bio**
+<!-- {/* DevList, devArray */} -->
+
+<!-- {/* TechList, shortStackArray */} -->
+Built with: React, Node, MongoDB, Tailwind
+<!-- {/* TechList, shortStackArray */} -->
+
+[nyt]: <https://www.nytimes.com/2023/08/02/business/media/meta-news-in-canada.html>
+[can]: <https://www.canada.ca/en/canadian-heritage/services/online-news.html>
+[insta]: <https://www.instagram.com/ExtraExtraCA>
+[website]: <https://www.extraextranews.ca>
+[resrc]: ('/resources')
+
+[linkedStub]: <https://www.linkedin.com/in/benjamin-stubina>
+[linkedGold]: <https://www.linkedin.com/in/bennygoldman>
+[gitStub]: <https://www.github.com/BenjaminStubina>
+[gitGold]: <https://www.github.com/bennygoldman>
