+++ Client/src/assets/content/resources.md
@@ -1,0 +1,14 @@
+# Resources
+
+This list is updated regularly to inform visitors of the latest developments to the Online News Act
+
+## Articles
+
+* "`title`", [source][articleUrl]. `date`
+
+## Podcasts
+
+* "`episode`", [podcast][episodeUrl]. `date`
+
+[articleUrl]: <http://www.xyz.com>
+[episodeUrl]: <http://www.xyz.com>
