+++ Client/src/hooks/useDialog.js
@@ -1,0 +1,11 @@
+import { useState } from "react";
+
+export function useDialog() {
+    const [isOpen, setIsOpen] = useState(false);
+
+    const toggleDialog = () => {
+        setIsOpen(!isOpen);
+    };
+
+    return { isOpen, toggleDialog };
+}
